@@ -1,5 +1,5 @@
 plugins {
-    id 'org.springframework.boot' version '2.7.5'
+    id 'org.springframework.boot' version '3.0.1'
     id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'java'
     id "me.champeau.jmh" version "0.6.4"
@@ -7,7 +7,7 @@ plugins {
 
 group = 'com.dalcho'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 configurations {
     compileOnly {
@@ -37,7 +37,7 @@ dependencies {
     implementation 'org.webjars:sockjs-client:1.1.2'
     implementation 'org.webjars:stomp-websocket:2.3.3-1'
     implementation 'org.webjars.bower:bootstrap:4.3.1'
-    implementation 'org.webjars:jquery:3.6.2'
+    implementation 'org.webjars:jquery:3.7.1'
 
     // chat-file
     implementation 'com.google.code.gson:gson:2.8.9'
@@ -71,8 +71,9 @@ dependencies {
     //implementation 'org.openjdk.jmh:jmh-core:1.32'
     //implementation 'org.openjdk.jmh:jmh-generator-annprocess:1.32'
 
-    // package org.slf4j does not exist
-    implementation 'org.slf4j:slf4j-api:1.7.36'
+    implementation 'org.slf4j:slf4j-simple:1.7.36'
+
+    implementation 'javax.annotation:javax.annotation-api:1.3.2'
 }
 jmh{
     fork = 1
