@@ -36,9 +36,9 @@ public class SecurityConfig {
 //					.httpBasic().disable() // Http basic Auth 기반으로 로그인 인증창이 열림(disable 시 인증창 열리지 않음)
 //					.exceptionHandling().authenticationEntryPoint(new RestAuthenticationEntryPoint())// 인증,인가가 되지 않은 요청 시 발생
 //					.and()
-        http.authorizeRequests()
-                .antMatchers("/css/**", "/oauth2/**", "/user/**", "/taste/**", "/js/**").permitAll()
-                .antMatchers("/admin/**").hasAuthority("ADMIN")
+        http.authorizeRequests().requestMatchers("/css/**", "/oauth2/**", "/user/**", "/taste/**", "/js/**")
+                .permitAll()
+                .requestMatchers("/admin/**").hasAuthority("ADMIN")
                 .anyRequest().authenticated();
 
         http.oauth2Login().loginPage("/user/login")
