@@ -1,15 +1,15 @@
 package com.dalcho.adme.config.security;
 
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @RequiredArgsConstructor
@@ -18,13 +18,9 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	private final JwtTokenProvider jwtTokenProvider;
 
 	@Override
-	public void doFilterInternal(
-			HttpServletRequest servletRequest,
-			HttpServletResponse servletResponse,
-			FilterChain filterChain) throws ServletException, IOException {
-
+	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
 		// servletRequest 에서 token 추출
-		String token = jwtTokenProvider.resolveToken(servletRequest);
+		String token = jwtTokenProvider.resolveToken(request);
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		log.info("[dofilterInternal] token 값 유효성 체크 시작");
 		try {
@@ -37,13 +33,13 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 			} else if (auth.getPrincipal() != null && token == null) {
 				log.info(" [  oauth  ]   login ");
 			} else {
-				log.info("[doFilterInternal] 유효한 JWT 토큰이 없습니다, uri: {}", servletRequest.getRequestURL());
+				log.info("[doFilterInternal] 유효한 JWT 토큰이 없습니다, uri: {}", request.getRequestURL());
 			}
 		} catch (NullPointerException e) {
 			log.info("doFilterInternal");
 		} finally {
 			// 요청 정보가 매칭되지 않을경우 동작
-			filterChain.doFilter(servletRequest, servletResponse);
+			filterChain.doFilter(request, response);
 		}
 	}
 }
