@@ -1,11 +1,14 @@
 package com.dalcho.adme.oauth2;
 
+import com.dalcho.adme.config.security.JwtTokenProvider;
 import com.dalcho.adme.exception.notfound.UserNotFoundException;
 import com.dalcho.adme.model.User;
-import com.dalcho.adme.config.security.JwtTokenProvider;
 import com.dalcho.adme.model.UserRole;
 import com.dalcho.adme.oauth2.util.UserMapper;
 import com.dalcho.adme.repository.UserRepository;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,8 +19,6 @@ import org.springframework.security.web.authentication.AuthenticationSuccessHand
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
@@ -32,9 +33,18 @@ public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
 	@Value("${oauth.admin.redirection.url}")
 	private String ADMIN_REDIRECTION_URL;
 
+	private String getRedirectionURI(String token, User user) {
+		if(user.getRole().name().equals(UserRole.USER.name())){
+			return UriComponentsBuilder.fromUriString(REDIRECTION_URL).queryParam("token", token).build().toUriString();
+		}else if (user.getRole().name().equals(UserRole.ADMIN.name())){
+			return UriComponentsBuilder.fromUriString(ADMIN_REDIRECTION_URL).queryParam("token", token).build().toUriString();
+		}else{
+			throw new RuntimeException("[error] 잘못된 권한입니다.");
+		}
+	};
+
 	@Override
-	public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws
-			IOException {
+	public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
 		OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
 		String email = (String) oAuth2User.getAttributes().get("email");
 		User userInfo = userRepository.findByEmail(email).orElseThrow(UserNotFoundException::new);
@@ -50,16 +60,6 @@ public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
 
 		response.sendRedirect(getRedirectionURI(token, user));
 	}
-
-	private String getRedirectionURI(String token, User user) {
-		if(user.getRole().name().equals(UserRole.USER.name())){
-			return UriComponentsBuilder.fromUriString(REDIRECTION_URL).queryParam("token", token).build().toUriString();
-		}else if (user.getRole().name().equals(UserRole.ADMIN.name())){
-			return UriComponentsBuilder.fromUriString(ADMIN_REDIRECTION_URL).queryParam("token", token).build().toUriString();
-		}else{
-			throw new RuntimeException("[error] 잘못된 권한입니다.");
-		}
-	};
 }
 
 
